Modified the buildmode of bitcode libraries.
The Default is Release+Asserts but if you are building for example "Release+Debug+Asserts" then this break because
KLEE will look for the bitcode libraries in the wrong place because the RUNTIME_CONFIGURATION macro is not defined
to be what KLEE actually builds as.

This has been tweaked so that when we build the bitcode libraries as "Release+Asserts" whether or not we've
choosen debug for the main build of KLEE does not matter.